requirements.txt
asgiref==3.9.1
coverage==7.9.2
Django==5.2.4
django-rest-framework==0.1.0
djangorestframework==3.16.0
pytz==2024.1
sqlparse==0.5.3
tzdata==2025.2
